requirements.txt
Flask==3.1.0
Flask-Login==0.6.3
Flask-Migrate==4.1.0
Flask-SQLAlchemy==3.1.1
Flask-WTF==1.2.2
gunicorn==23.0.0
python-dotenv==1.0.1
Jinja2==3.1.5
Werkzeug==3.1.3
WTForms==3.2.1
SQLAlchemy==2.0.40
itsdangerous==2.2.0
click==8.1.8
MarkupSafe==3.0.2
pyodbc==5.1.0
